requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-extensions==0.8.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
whitenoise==6.7.0
urllib3==2.5.0
PyYAML==6.0.2
inflection==0.5.1
Jinja2==3.1.6
django-jazzmin==3.0.1
django-admin-rangefilter==0.13.3
django-chartjs==2.3.0
django-reversion==6.0.0
ruff==0.13.2
stripe==9.6.0
